+++ setupTests.js
@@ -1,6 +1,7 @@
 import '@testing-library/jest-dom'
 import dotenvFlow from 'dotenv-flow'
 import { TextEncoder, TextDecoder } from 'util'
+import { NotifyClient } from 'notifications-node-client'
 
 global.TextEncoder = TextEncoder
 global.TextDecoder = TextDecoder
@@ -17,4 +18,34 @@ jest.mock('@sentry/nextjs', () => {
     setUser: jest.fn(() => 'Sentry.setUser'),
     setTag: jest.fn(() => 'Sentry.setTag'),
   }
+})
+
+jest.mock('axios', () => {
+  const mockClient = {
+    get: jest.fn(),
+    post: jest.fn(),
+  }
+
+  return {
+    default: {
+      create: jest.fn(() => mockClient),
+    },
+    create: jest.fn(() => mockClient),
+    __mockClient: mockClient,
+  }
+})
+
+jest.mock('notifications-node-client')
+
+beforeAll(() => {
+  const mockSendEmail = jest.fn().mockResolvedValue({
+    data: { message: 'Email sent' }
+  })
+
+  const mockPrepareUpload = jest.fn().mockReturnValue('mocked-file-upload')
+
+  NotifyClient.mockImplementation(() => ({
+    sendEmail: mockSendEmail,
+    prepareUpload: mockPrepareUpload
+  }))
 })
